Fix gba_cart.ld support for overlays (#4)
* Fix gba_cart.ld support for overlays

The current linker script uses `__iwram_overlay_end - __iwram_overlay_start` as an offset for `__ewram_lma`, which uses only the maximum size of an IWRAM overlay and is wrong even for only 2 overlays (section `.ewram`'s LMA will overlap with `.iwram1`'s LMA). This fix makes it correctly consider the total size of all overlays when computing the LMA for section `.ewram`.

* Remove overlay references on main section

The original section definition for `.iwram` includes all sections that start with `.iwram` on it, which means code supposed to go in IWRAM overlays would be placed on the `.iwram` definition, rendering the overlay feature useless. This commit fix this patch.